backend/requirements.txt
django==5.2.4
markdown
django-environ 
django-filter
django-leaflet
djangorestframework
django-cors-headers
ipython
pytest
pytest-django
requests
beautifulsoup4
psycopg2-binary
drf-spectacular
python-dotenv
gunicorn
uvicorn
pillow
psycopg[binary]
djangorestframework-simplejwt
drf_social_oauth2==3.1.0
tqdm
pandas
dj-rest-auth==7.0.1
django-allauth==0.63.6
elasticsearch==8.17.0
elasticsearch-dsl==8.17.0
django-elasticsearch-dsl==8.0
